refactor(desktop): extract the command switch into commandRunner (F5.5)
executeCommand's ~490-line dispatch switch moves verbatim into
runCommand(input, ctx) (commandRunner.ts), with the ~90 values/handlers
it touches passed as a typed CommandCtx (commandCtx.ts). App's
executeCommand is now a thin wrapper building the ctx. The switch body is
byte-identical — behavior preserved by construction.

App.tsx 3,474 -> 3,066; commandRunner 454, commandCtx 109 (both < 500).
unit 61 / e2e 36 green (every command flow runs through runCommand),
tsc + build clean.
